refactor: switch to ES module boot
