+++ index.html
@@ -800,7 +800,6 @@ self.onmessage = function(e) {
   }
 };
 </script>
-<script defer src="assets/js/state/store.js"></script>
-<script defer src="assets/js/app.js"></script>
+<script type="module" src="assets/js/boot.js"></script>
 </body>
 </html>
